requirements.txt
alembic==1.11.1
anyio==3.7.0
beautifulsoup4==4.12.2
black==23.3.0
bs4==0.0.1
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
ecdsa==0.18.0
exceptiongroup==1.1.1
fastapi==0.98.0
greenlet==2.0.2
h11==0.14.0
idna==3.4
Levenshtein==0.21.1
Mako==1.2.4
MarkupSafe==2.1.3
mypy-extensions==1.0.0
mysql-connector==2.2.9
numpy==1.25.0
packaging==23.1
pandas==2.0.2
passlib==1.7.4
pathspec==0.11.1
Pillow==9.5.0
platformdirs==3.8.0
pyasn1==0.5.0
pydantic==1.10.9
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
python-Levenshtein==0.21.1
pytz==2023.3
rapidfuzz==3.1.1
requests==2.31.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
SQLAlchemy==2.0.16
starlette==0.27.0
tomli==2.0.1
tqdm==4.65.0
typing_extensions==4.6.3
tzdata==2023.3
urllib3==2.0.3
uvicorn==0.22.0
